+++ examples/recv-iterate-ts.ts
@@ -5,19 +5,25 @@ import {DgramAsPromised} from "../src/dgram-as-promised.js"
 const port = Number(process.argv[2]) || 0
 
 async function main(): Promise<void> {
-  const socket = DgramAsPromised.createSocket("udp6")
+  const socket = DgramAsPromised.createSocket("udp4")
 
-  const address = await socket.bind({port, address: "::1"})
+  const address = await socket.bind({port})
   console.info(`Socket is listening on ${address.address}:${address.port}`)
+  socket.setTimeout(5000)
 
-  for await (const packet of socket) {
-    console.log(packet)
-    if (packet.msg.indexOf(4) !== -1) {
-      await socket.close()
+  try {
+    for await (const packet of socket) {
+      console.log(packet)
+      if (packet.msg.indexOf(4) !== -1) {
+        await socket.close()
+      }
     }
+  } catch (e) {
+    console.error(e)
+  } finally {
+    await socket.close()
+    socket.destroy()
   }
-
-  socket.destroy()
 }
 
 main().catch(console.error)