+++ examples/recv-iterate.js
@@ -1,23 +1,29 @@
 #!/usr/bin/env node
 
+import wtf from "wtfnode"
+
 import {DgramAsPromised} from "../lib/dgram-as-promised.js"
 
 const port = Number(process.argv[2]) || 0
 
 async function main() {
-  const socket = DgramAsPromised.createSocket("udp4")
+  const socket = DgramAsPromised.createSocket("udp6")
 
-  const address = await socket.bind({port})
+  const address = await socket.bind({port, address: "::1"})
   console.info(`Socket is listening on ${address.address}:${address.port}`)
+  socket.setTimeout(5000)
 
-  for await (const packet of socket) {
-    console.log(packet)
-    if (packet.msg.indexOf(4) !== -1) {
-      await socket.close()
+  try {
+    for await (const packet of socket) {
+      console.log(packet)
+      if (packet.msg.indexOf(4) !== -1) {
+        await socket.close()
+      }
     }
+  } catch (e) {
+    console.error(e)
   }
-
-  socket.destroy()
+  wtf.dump()
 }
 
 main().catch(console.error)