+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   `setRecvBufferSize`, `setSendBufferSize` and `destroy` return `this`.
 - `address` method doesn't return `string` anymore; closes:
   dex4er/js-dgram-as-promised#51
+- `setTimeout` method sets timeout for `recv` method; closes:
+  dex4er/js-dgram-as-promised#55
 
 ## v5.0.1 2020-10-16